Move snapshot store funcs to pkg
